requirements.txt
psycopg2-binary
pandas
Flask-SQLAlchemy